removed GET requests from service_shifts, added it to new file schedule_get because schedule.py already exists containing the POST logic. also added routes to app.py. POST and GET branches are now merged
